requirements.txt
Django
gunicorn
django-environ
djangorestframework-simplejwt
djangorestframework
django-debug-toolbar
drf_spectacular
psycopg2-binary
django-filter
django-redis
celery
